controller: add key up/down/sync/reset controls
Added key up, down, sync and reset button functionality.

Note: the up and down key button LED's appear to be automatic, ie: cannot be programmed to light up steady to show the current direction of pitch change.  Automatic behaviour is to provide feedback indicating that the button is being pressed.